refactor MCP client into cohesive modules
